Refactor: DashboardResponse 및Dashboard types
