+++ src/types/dashboard.ts
@@ -1,33 +1,12 @@
-export interface ColumnList {
-  id: number;
-  title: string;
-  teamId: string;
-  dashboardId: number;
-  createdAt: string;
-  updatedAt: string;
-}
+import { ColumnListResponse, CardListResponse } from '../apis/schema/dashboardResponse';
+
+export type ColumnList = ColumnListResponse['data'][number];
 
 export interface CardList {
-  cards: Cards[];
+  cards: Cards;
   totalCount: number;
   cursorId: number;
 }
 
-export interface Cards {
-  id: number;
-  title: string;
-  description: string;
-  tags: string[];
-  dueDate: string;
-  assignee?: {
-    id: number;
-    nickname: string;
-    profileImageUrl?: string;
-  };
-  imageUrl?: string;
-  teamId: string;
-  dashboardId: number;
-  columnId: number;
-  createdAt: string;
-  updatedAt: string;
-}
+export type Cards = CardListResponse['cards'][number];
+export type Assignee = Cards['assignee'];
